pyproject.toml
[tool.poetry]
name = "pokepoke_miss"
version = "0.1.16"
description = "wmc的戳一戳回复插件，消息内容、概率都可以定制喔~"
authors = ["Miaowing <shengwang52005@163.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/shengwang52005/pokepoke_miss"
repository = "https://github.com/shengwang52005/pokepoke_miss"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = ">=2.3.2"
nonebot-adapter-onebot = ">=2.3.1"
httpx = ">=0.27.0"
pil_utils = "0.1.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
